Fix <Icon /> hydration error
Icons were <div>s, which, when used in MDX, nested in <p>s, which is
invalid and throws an error/warning from React.

Changing them to <span>s is good and semantic but also necessitated
changes to the selectors in PointSelector, which was apparently
incorrectly using :first-of-type.

All of this is annoying and makes for more code but is more explicit
so probably better?!